+++ wasm_proto/core/mem.h
@@ -2,20 +2,20 @@
  * `mem.h` declares the LC-3 memory map
  */
 
-#include "uint.h"
+#include <stdint.h>
 
-extern uint *trp_vect_tbl;
-extern uint trp_vect_tbl_limit;
+extern uint16_t *trp_vect_tbl;
+extern uint16_t trp_vect_tbl_limit;
 
-extern uint *int_vect_tbl;
-extern uint int_vect_tbl_limit;
+extern uint16_t *int_vect_tbl;
+extern uint16_t int_vect_tbl_limit;
 
-extern uint *sup_stack;
-extern uint sup_stack_limit;
+extern uint16_t *sup_stack;
+extern uint16_t sup_stack_limit;
 
-extern uint *usr_stack;
-extern uint sup_stack_limit;
+extern uint16_t *usr_stack;
+extern uint16_t sup_stack_limit;
 
-extern uint *io_dev;
-extern uint io_dev_limit;
+extern uint16_t *io_dev;
+extern uint16_t io_dev_limit;
