+++ wasm_proto/core/subr.h
@@ -2,20 +2,20 @@
  * `subr.h` declares all LC-3 subroutines
  */
 
-#include "uint.h"
+#include <stdint.h>
 
-extern void and(uint i);
-extern void not(uint i);
-extern void br(uint i);
-extern void st(uint i);
-extern void ld(uint i);
-extern void str(uint i);
-extern void ldr(uint i);
-extern void jmp(uint i);
-extern void jsr(uint i);
-extern void ldi(uint i);
-extern void sti(uint i);
-extern void lea(uint i);
-extern void trap(uint i);
-extern void rti(uint i);
+extern void and(uint16_t i);
+extern void not(uint16_t i);
+extern void br(uint16_t i);
+extern void st(uint16_t i);
+extern void ld(uint16_t i);
+extern void str(uint16_t i);
+extern void ldr(uint16_t i);
+extern void jmp(uint16_t i);
+extern void jsr(uint16_t i);
+extern void ldi(uint16_t i);
+extern void sti(uint16_t i);
+extern void lea(uint16_t i);
+extern void trap(uint16_t i);
+extern void rti(uint16_t i);
