+++ wasm_proto/core/wasm.c
@@ -2,10 +2,10 @@
  * `wasm.c` handles Wasm-specific tasks
  */
 
-//#include <emscripten/emscripten.h>
+#include <stdint.h>
+#include <emscripten/emscripten.h>
 #include "core.h"
 #include "mem.h"
-#include "uint.h"
 
 // EMSCRIPTEN_KEEPALIVE
 void init() {
@@ -17,32 +17,32 @@ void init() {
 }
 
 // EMSCRIPTEN_KEEPALIVE
-uint reg(int i) {
+uint16_t reg(uint16_t i) {
 	return core.gpr[i];
 }
 
 // EMSCRIPTEN_KEEPALIVE
-uint pc() {
+uint16_t pc() {
 	return core.pc;
 }
 
 // EMSCRIPTEN_KEEPALIVE
-uint privilege() {
+uint16_t privilege() {
 	return core.psr >> 15 & 0x1;
 }
 
 // EMSCRIPTEN_KEEPALIVE
-uint priority() {
+uint16_t priority() {
 	return core.psr >> 8 & 0x7;
 }
 
 // EMSCRIPTEN_KEEPALIVE
-uint* pass1(uint *src);
+uint16_t* pass1(uint16_t *src);
 	// handles labels and assembler directives
 	// returns pointer to program in memory for second pass
 	//    i.e. pass2(pass1( ... ));
 
 // EMSCRIPTEN_KEEPALIVE
-void pass2(uint *src);
+void pass2(uint16_t *src);
 	// read at `src` and execute intructions
